Move checks from directories to files
