+++ colin/checks/dockerfile.py
@@ -1,12 +1,13 @@
-from colin.checks.abstract.dockerfile import DockerfileCheck
+from colin.checks.abstract.dockerfile import (DockerfileCheck,
+                                              InstructionCountCheck)
 from colin.checks.result import CheckResult
 from colin.core.target import ImageName
 
 
-class FromTagCheck(DockerfileCheck):
+class FromTagNotLatestCheck(DockerfileCheck):
 
     def __init__(self):
-        super(FromTagCheck, self) \
+        super(FromTagNotLatestCheck, self) \
             .__init__(name="from_tag_not_latest",
                       message="In FROM, tag has to be specified and not 'latest'.",
                       description="Using the 'latest' tag may cause unpredictable builds."
@@ -24,3 +25,16 @@ class FromTagCheck(DockerfileCheck):
                            reference_url=self.reference_url,
                            check_name=self.name,
                            logs=[])
+
+
+class MaintainerDeprecatedCheck(InstructionCountCheck):
+
+    def __init__(self):
+        super(MaintainerDeprecatedCheck, self) \
+            .__init__(name="maintainer_deprecated",
+                      message="Dockerfile instruction `MAINTAINER` is deprecated.",
+                      description="Replace with label 'maintainer'.",
+                      reference_url="https://docs.docker.com/engine/reference/builder/#maintainer-deprecated",
+                      tags=["maintainer", "dockerfile", "deprecated"],
+                      instruction="MAINTAINER",
+                      max_count=0)
