+++ colin/checks/dynamic.py
@@ -1,11 +1,11 @@
 from colin.checks.abstract.cmd import CmdCheck
 
 
-class BashCheck(CmdCheck):
+class ShellRunableCheck(CmdCheck):
 
     def __init__(self):
-        super(BashCheck, self) \
-            .__init__(name="sh",
+        super(ShellRunableCheck, self) \
+            .__init__(name="shell_runnable",
                       message="Shell has to be runnable.",
                       description="The target has to be able to invoke shell.",
                       reference_url="https://docs.docker.com/engine/reference/run/",
