--- colin/checks/labels/uninstall_capital_deprecated.py
@@ -1,14 +1,0 @@
-from colin.checks.abstract.labels import DeprecatedLabelCheck
-
-
-class UninstallLabelCapitalDeprecatedCheck(DeprecatedLabelCheck):
-
-    def __init__(self):
-        super(UninstallLabelCapitalDeprecatedCheck, self) \
-            .__init__(name="uninstall_label_capital_deprecated",
-                      message="Label 'UNINSTALL' is deprecated.",
-                      description="Replace with 'uninstall'.",
-                      reference_url="?????",
-                      tags=["uninstall", "label", "capital", "deprecated"],
-                      old_label="UNINSTALL",
-                      new_label="uninstall")
